Streamline generating list of filesystems; handle backward compatibility
Implement logic in a single location for reading the filesystem source
and target devices out of the config file and creating a generator
which iterates through them all, and use that function whenever a list
of filesystems is required.

Handle transparent backward compatibility with the old version of the
script that didn't store filesystem targets in the config file and
only used one key instead of two.

Be slightly more verbose about when we are adding and removing keys so
that it doesn't seem to the user like the script is hung when it's
working.

Make the messages about adding and removing keys display both the
target and source for the filesystem, not just the target.