Add --default-kernel and --kernel-version to enable-{once,always}
Allow the user to indicate that only the default kernel should be
configured for keyless entry by specifying `--default-kernel` with
`enable-once` or `enable-always`, or to specify one or more specific
kernel versions to enable for with `--kernel-version`. This makes
enable and disable operations run faster by only rebuilding a subset
of the initial RAM disks.