Disable kernel versions added after we were enabled
If we are enabled for specific kernel versions and then new kernels
are installed while we are enabled, we'll be enabled for those
versions as well.

Therefore, rather than storing in the config the specific kernel
versions for which we were enabled, store the kernel versions for
which we were _not_ enabled, and then when disabling, rebuild the
initramfs for all the ones not explicitly listed as skipped.