Better fix for selective kernel versions
The fix in 1722aa for making sure initrds created after we were
enabled are regenerated when we are disabled is inadequate. The
problem it doesn't address is that initrds for existing kernel
versions can get rebuilt for a number of reasons other than a new
kernel version being enabled.

The correct fix for this is to record the timestamp when we enable and
then, when disabling, rebuild any initrds with timestamps newer than
when we enabled.

Fixes #5